# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'ModWorkshop API Documentation'
  description: "The ModWorkshop API is available for everyone to use. \r\n\r\nDo note that by using the API you must follow the following guidelines:\r\n\r\n1. Do not spam the API.\r\n\r\n2. Do not replicate the site or remove the need to visit the site to download mods. \r\nExceptions: You are allowed to implement updates, you are allowed to make a mod downloadable to ease the process of joining a game with mods (Example: downloading maps). Integration to the site itself is fine too (We'll add ways to download mods directly to mod managers soon).\r\n\r\n3. Respect user's privacy and do not store their information without their consent.\r\n\r\nAt the moment, the API <b>only supports GET requests</b>. More support will come in the future, but will require the use of API keys."
  version: 1.0.0
servers:
  -
    url: 'https://api.modworkshop.net'
paths:
...
  '/mods/{mod_id}/toggle-liked':
    post:
      summary: 'Toggle like'
      operationId: toggleLike
      description: 'Toggles the state of the like of the mod'
      parameters: []
      responses: {  }
      tags:
        - Mods
    parameters:
      -
        in: path
        name: mod_id
        description: 'The ID of the mod.'
        example: 1
        required: true
        schema:
          type: integer
...